computer-use-demo/computer_use_demo/requirements.txt
streamlit==1.41.0
anthropic[bedrock,vertex]>=0.39.0
jsonschema==4.22.0
boto3>=1.28.57
google-auth<3,>=2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
